+++ ProgAvanceeProjectWPF/DAO/MemberDAO.cs
@@ -152,7 +152,7 @@ internal class MemberDAO : DAO<Member>
                             reader.GetString("telephone"),
                             reader.GetString("login"),
                             reader.GetString("password"),
-                            reader.GetFloat("balance")
+                            reader.GetDouble("balance")
                             );
                     }
                 }
@@ -187,7 +187,7 @@ internal class MemberDAO : DAO<Member>
                             reader.GetString("telephone"),
                             reader.GetString("login"),
                             reader.GetString("password"),
-                            reader.GetFloat("balance")
+                            reader.GetDouble("balance")
                             );
                         member.Categories = categoryDAO.FindAllByMember(member);
                         member.Bikes = bikeDAO.FindAllByMember(member);
